+++ src/panel/RemoteEditPanelHandlerTypes.ts
@@ -21,6 +21,7 @@ export interface RemoteEditPanelMessageHandlers {
   requestCreateFile(payload: any): Promise<void>;
   requestCreateDirectory(payload: any): Promise<void>;
   requestMakeCopy(payload: any): Promise<void>;
+  requestCalculateChecksums(payload: any): Promise<void>;
   requestRenameEntry(payload: any): Promise<void>;
   requestDeleteEntry(payload: any): Promise<void>;
   requestDeleteEntries(payload: any): Promise<void>;